requirements.txt

# Core Web Framework
streamlit>=1.28.0

# Data Processing & Analysis
pandas>=2.0.0
numpy>=1.24.0

# Financial Data
yfinance>=0.2.0

# HTTP Requests
requests>=2.31.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Document Processing (Optional but recommended)
python-docx>=1.1.0
openpyxl>=3.1.0

# Text Processing
nltk>=3.8.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# Database Clients (Stubs - install if using actual services)
# pinecone-client>=2.2.0
# neo4j>=5.12.0

# Environment Variables
python-dotenv>=1.0.0

# Utilities
pathlib>=1.0.1
typing-extensions>=4.8.0

# Testing (Optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Code Quality (Optional)
black>=23.9.0
flake8>=6.1.0